% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBook.R
\name{buildBook}
\alias{buildBook}
\title{Build a Book without Underscores}
\usage{
buildBook(output_dir = "book", ...)
}
\arguments{
\item{output_dir}{Output directory as \code{character}.}

\item{...}{Arguments passed to \code{\link[bookdown]{render_book}}.}
}
\description{
Since the use of underscores ('_') is not permitted when streaming 
\strong{bookdown} documents via 
\href{https://pages.github.com/}{GitHub Pages}, this wrapper function serves 
to remove any unwanted underscores from subfolders and link \code{.html} 
documents created by \code{\link[bookdown]{render_book}}.
}
\note{
While all remaining arguments passed to \code{\link[bookdown]{render_book}} 
remain untouched, and hence, their specification is freely up to the user, 
the default value of 'output_dir' is explicitly set to \code{"book"} here. If 
this were not the case (i.e. if the default value were used), the output 
document would be created in \code{"_book"} which is not desirable for 
obvious reasons.
}
\seealso{
\code{\link[bookdown]{render_book}}.
}
\author{
Florian Detsch
}
