% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOffsetGridText.R
\name{calcOffsetGridText}
\alias{calcOffsetGridText}
\title{Calculate offset coordinates for grid-based text annotations}
\usage{
calcOffsetGridText(x, y = NULL, offset = 0.02, pos = NULL, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{x}{A \code{numeric} vector containing x coordinates, or a 2-column
\code{matrix} containing x and y coordinates.}

\item{y}{A \code{numeric} vector containing y coordinates, or \code{NULL} 
if 'x' is a two-column \code{matrix}.}

\item{offset}{A \code{numeric} offset in normalized parent coordinates
("npc", see \code{\link[grid]{unit}}).}

\item{pos}{Text position specifier(s) as \code{integer} used by 
\code{\link{text}}. If not supplied, optimal text positions will be 
determined with respect to neighboring locations using 
\code{\link[plotrix]{thigmophobe}}.}

\item{xlim, ylim}{X and Y-axis limits (\code{c(min, max)}) of the current plot. 
If not supplied, limits are automatically calculated from supplied x and y
coordinates.}

\item{...}{Currently not used.}
}
\value{
A \code{numeric matrix} containing offset coordinates.
}
\description{
Calculate offset coordinates for (\strong{grid}-based) text drawing 
functions, e.g. \code{\link{grid.text}}.
}
\examples{
stopifnot(
  require(mapview)
  , require(lattice)
  , require(grid)
)

## calculate offsets for breweries with more than 3 different types of beer
brw = as(subset(breweries, number.of.types > 3), "Spatial")
loc = calcOffsetGridText(coordinates(brw), offset = .025)

## create plot
p = spplot(brw, zcol = "number.of.types", auto.key = FALSE)

plot.new()
print(p, newpage = FALSE)

## add text labels
downViewport(trellis.vpname(name = "figure"))
for (i in 1:length(brw)) {
  grid.text(label = brw$number.of.types[i], x = loc[i, 1], y = loc[i, 2])
}
              
}
\seealso{
\code{\link[grid]{grid.text}}, \code{\link{text}}, 
\code{\link[plotrix]{thigmophobe}}
}
\author{
Florian Detsch
}
