% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext2spy.R
\name{ext2spy}
\alias{ext2spy}
\title{Convert Spatial Extent to Polygon}
\usage{
ext2spy(x, crs = "+init=epsg:4326")
}
\arguments{
\item{x}{An \code{Extent} object, or any object from which an \code{Extent} 
can be extracted, eg. \code{Raster*}.}

\item{crs}{Coordinate reference system passed to \code{\link{proj4string}}.}
}
\value{
A \code{SpatialPolygons} object.
}
\description{
Convert a spatial extent to polygons.
}
\examples{
ext = extent(c(25, 70, -5, 30))

}
\seealso{
\code{\link{extent}}.
}
\author{
Florian Detsch
}
