% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orcs-package.R
\docType{data}
\name{KiLi}
\alias{KiLi}
\title{Bing Aerial Image of Kilimanjaro}
\format{A \code{"RasterStack-class"} with 3 bands (red, green, blue).}
\description{
Bing aerial image of Kilimanjaro downloaded from 
\href{https://www.openstreetmap.org/}{OpenStreetMap}.
}
\details{
Copyright: OpenStreetMap contributors, see 
\url{https://www.openstreetmap.org/copyright}.
}
