% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knit2md.R
\name{knit2md}
\alias{knit2md}
\title{Convert a R Markdown file to ordinary Markdown}
\usage{
knit2md(path_in = getwd(), path_out = NULL, pttrn = "*.Rmd$", ...)
}
\arguments{
\item{path_in}{Character. Input file path; defaults to the current
working directory.}

\item{path_out}{Character. Output file path; if \code{NULL},
defaults to \code{path_in}.}

\item{pttrn}{Character. A regular expression as taken by
\code{\link{list.files}}; defaults to \code{"*.Rmd$"}.}

\item{...}{Further arguments passed on to \code{\link{knit}}.}
}
\value{
Output filename(s).
}
\description{
This function is a convenient wrapper around \code{\link{knit}} as
it automatically converts all R Markdown files (*.Rmd) located in
a specified folder and, optionally, matching a particular pattern
to standard Markdown (*.md).
}
\author{
Florian Detsch
}
\seealso{
\code{\link{list.files}}, \code{\link{knit}}
}

