% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lineEnding.R
\name{lineEnding}
\alias{lineEnding}
\title{Convert between DOS and UNIX line endings}
\usage{
lineEnding(infile, outfile = NULL, to = c("dos", "unix"), ...)
}
\arguments{
\item{infile}{Input filename.}

\item{outfile}{Output filename. If not supplied, 'infile' will be
overwritten.}

\item{to}{Either 'dos' or 'unix'.}

\item{...}{Further arguments passed on to \code{\link{system}}.}
}
\description{
This function converts between DOS and UNIX style line endings by envoking
\code{unix2dos} (or \code{dos2unix}) upon a text file. See also
\code{system("unix2dos --help")}.
}
\examples{
## input file
infile <- paste(system.file(package = "Orcs"), "DESCRIPTION", sep = "/")

## convert to dos line endings and write to output file
lineEnding(infile, outfile = "~/Desktop/DESCRIPTION4wd", to = "dos")
}
\author{
Florian Detsch
}

