% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind2df.R
\name{list2df}
\alias{list2df}
\title{Create \code{data.frame} from \code{list}}
\usage{
list2df(x, ...)
}
\arguments{
\item{x}{A \code{list} object.}

\item{...}{Further arguments passed on to \code{\link{data.frame}}.}
}
\value{
A \code{data.frame} object.
}
\description{
Create a \code{data.frame} from a \code{list} directly, \emph{i.e.} without 
being required to explicitly call \code{\link{rbind}} first.
}
\examples{
lst <- list(letters[1:3], letters[4:6], letters[7:9])

do.call("rbind", lst) # results in matrix
list2df(lst)          # results in data.frame

}
\seealso{
\code{\link{data.frame}}, \code{\link{rbind}}, \code{\link{do.call}}.
}

