% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPkgs.R
\name{loadPkgs}
\alias{loadPkgs}
\title{Load multiple packages}
\usage{
loadPkgs(pkgs)
}
\arguments{
\item{pkgs}{'character'. Packages to load.}
}
\description{
Being a wrapper around \code{\link{library}}, this function allows you to 
load and attach multiple packages at once.
}
\examples{
loadPkgs(c("raster", "rgdal"))

}
\author{
Florian Detsch
}
\seealso{
\code{\link{library}}.
}

