% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanDifference.R
\docType{methods}
\name{meanDifference}
\alias{meanDifference}
\alias{meanDifference,RasterLayer-method}
\alias{meanDifference,numeric-method}
\title{Calculate mean difference between two datasets}
\usage{
\S4method{meanDifference}{RasterLayer}(x, y)

\S4method{meanDifference}{numeric}(x, y)
}
\arguments{
\item{x,y}{Objects of class 'RasterLayer' or 'numeric' vectors.}
}
\value{
The mean difference between values of 'x' and 'y'.
}
\description{
Calculate the mean difference between two datasets as suggested by
\emph{Wang et al. (2012)}. This function can be operated either with 'numeric'
vectors or objects of class 'RasterLayer'.


}
\examples{
x <- 1:10
y <- 2:11
meanDifference(x, y)
}

