% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\alias{merge,list-method}
\title{Merge Objects Stored in a List}
\arguments{
\item{x}{A \code{list} of objects of the same type (eg \code{Raster*}).}

\item{...}{Additional arguments passed to the underlying merge method (eg 
arguments compatible with \code{\link[raster]{merge}} and 
\code{\link[raster]{writeRaster}} for \code{Raster*} input).}
}
\value{
A merged object (eg a new \code{Raster*} object with a larger spatial 
extent).
}
\description{
Complementing existing merge methods, eg \code{\link[raster]{merge}} for 
\code{Raster*} objects, which typically work with one or two inputs only, 
this function accepts a \code{list} of objects that are to be merged 
together.
}
\examples{
## Raster* input
dms = list.files(system.file("extdata", package = "Orcs")
                 , pattern = "ASTGTM2.*dem.tif$", full.names = TRUE)
dms = lapply(dms, raster)

dem = merge(dms, tolerance = 1e4)                  
plot(dem)

}
\seealso{
\code{\link{merge}}, \code{\link{do.call}}.
}
\author{
Florian Detsch
}
