% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiKnit.R
\name{multiKnit}
\alias{multiKnit}
\title{Convert multiple R Markdown files to ordinary Markdown}
\usage{
multiKnit(path_in = getwd(), path_out = NULL, pttrn = "*.Rmd$", ...)
}
\arguments{
\item{path_in}{Character. Input file path; defaults to the current
working directory.}

\item{path_out}{Character. Output file path; if \code{NULL},
defaults to \code{path_in}.}

\item{pttrn}{Character. A regular expression as taken by
\code{\link{list.files}}; defaults to \code{"*.Rmd$"}.}

\item{...}{Further arguments passed on to \code{\link{knit}}.}
}
\value{
Output filenames.
}
\description{
This function is a convenient wrapper around \code{\link{knit}} as
it automatically converts multiple R Markdown files (*.Rmd) located in
a specified folder (and, optionally, matching a particular pattern)
to standard Markdown (*.md).
}
\author{
Florian Detsch
}
\seealso{
\code{\link{list.files}}, \code{\link{knit}}
}

