% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetGridText.R
\name{offsetGridText}
\alias{offsetGridText}
\title{Insert offset text annotation into 'trellis' plot}
\usage{
offsetGridText(x, y = NULL, labels, xlim = NULL, ylim = NULL,
  pos = NULL, stext = FALSE, offset = 0.02, ...)
}
\arguments{
\item{x}{A \code{numeric} vector containing x coordinates, or a 2-column
\code{matrix} containing x and y coordinates.}

\item{y}{A \code{numeric} vector containing y coordinates, or \code{NULL} 
if 'x' is a two-column \code{matrix}.}

\item{labels}{The text to be written as \code{character}.}

\item{xlim, ylim}{X and Y-axis limits (\code{c(min, max)}) of the current plot. 
If not supplied, limits are automatically calculated from supplied x and y
coordinates.}

\item{pos}{Text position specifier(s) as \code{integer} used by 
\code{\link{text}}. If not supplied, optimal text positions will be 
determined with respect to neighboring locations using 
\code{\link[plotrix]{thigmophobe}}.}

\item{stext}{\code{logical}, defaults to \code{FALSE}. If \code{TRUE}, shadow 
text will be drawn around 'labels'.}

\item{offset}{A \code{numeric} offset in normalized parent coordinates
("npc", see \code{\link[grid]{unit}}).}

\item{...}{Additional arguments passed to the respective \strong{grid} text 
drawing function (depends on 'stext').}
}
\description{
This is a wrapper function around \code{\link{calcOffsetGridText}} and 
\strong{grid}-based text drawing functions (currently including 
\code{\link{grid.text}} and \code{\link{grid.stext}}) that automatically adds
offset text annotations to a 'trellis' plot.
}
\examples{
\dontrun{
#' bing satellite image of mt. kilimanjaro
rst_kili <- kiliAerial(minNumTiles = 12,
                       projection = "+init=epsg:4326")
spl_kili <- rgb2spLayout(rst_kili, alpha = .8)

#' kilimanjaro peaks
kibo <- c(-3.065053, 37.359031)
mawenzi <- c(-3.095436, 37.455061)
shira <- c(-3.038222, 37.210408)

mat_peaks <- rbind(kibo, mawenzi, shira)
df_peaks <- data.frame(peak = c("Kibo", "Mawenzi", "Shira"), 
                       x = mat_peaks[, 2], 
                       y = mat_peaks[, 1])

coordinates(df_peaks) <- ~ x + y
projection(df_peaks) <- "+init=epsg:4326"

#' visualization
library(latticeExtra)

xlim_kili <- c(37.15, 37.55)
ylim_kili <- c(-3.25, -2.9)

p_kili <- spplot(df_peaks, auto.key = FALSE, col.regions = "white", 
                 xlim = xlim_kili, ylim = ylim_kili, cex = 2, pch = 20,
                 scales = list(draw = TRUE), sp.layout = spl_kili) + 
  layer(sp.points(df_peaks, cex = 1.5, pch = 20, col = "black"))

print(p_kili)

library(grid)
downViewport(trellis.vpname(name = "figure"))
offsetGridText(x = coordinates(df_peaks), labels = c("Kibo", "Mawenzi", "Shira"),  
               xlim = xlim_kili, ylim = ylim_kili, stext = TRUE, offset = .02,
               gp = gpar(fontsize = 20, fontfamily = "Bookman Old Style"))
}
                               
}
\seealso{
\code{\link[grid]{grid.text}}, \code{\link{grid.stext}}, 
\code{\link[plotrix]{thigmophobe}}, \code{\link{calcOffsetGridText}}.
}
\author{
Florian Detsch
}
