\name{Orcs-package}
\alias{Orcs-package}
\alias{Orcs}
\docType{package}
\title{
\packageTitle{Orcs}
}
\description{
\packageDescription{Orcs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Orcs}
\packageIndices{Orcs}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Orcs}

Maintainer: \packageMaintainer{Orcs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
