% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par7zip.R
\name{par7zip}
\alias{par7zip}
\title{Parallelized 7zip-based compression}
\usage{
par7zip(path_in = getwd(), nodes = 1L, ...)
}
\arguments{
\item{path_in}{Character. Input file path; defaults to the current
working directory.}

\item{nodes}{Integer. Number of cores to use for parallelization; defaults to 
\code{1}.}

\item{...}{Further arguments passed on to \code{\link{list.files}}.}
}
\value{
Output filename(s).
}
\description{
By calling the Terminal (UNIX-based OS) or Command Prompt (Windows OS) from 
within R, this function performs parallelized 7zip-based compression of 
selected files based on the built-in \strong{parallel} package.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{list.files}}, \code{\link{system}}
}

