% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pureBasename.R
\name{pureBasename}
\alias{pureBasename}
\title{Return filename without extension}
\usage{
pureBasename(path)
}
\arguments{
\item{path}{Filename; see \code{\link{basename}}.}
}
\description{
As opposed to \code{\link{basename}}, this function returns the pure basename
of a file path, i.e. without extension.
}
\examples{
ch_file <- "path/to/file.with.extension"
pureBasename(ch_file)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{basename}}
}

