% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDuplCols.R
\name{rmDuplCols}
\alias{rmDuplCols}
\title{Remove duplicated columns from data frame}
\usage{
rmDuplCols(x, keep_first = TRUE, ...)
}
\arguments{
\item{x}{\code{data.frame}. Input data set to be processed.}

\item{keep_first}{Logical. Determines whether the first column of an otherwise 
duplicated header should be kept; defaults to \code{TRUE}.}

\item{...}{Currently not in use.}
}
\value{
Revised \code{data.frame}.
}
\description{
This function automatically detects and removes columns from a data frame 
based on duplicated headers.
}
\examples{
## sample data
mat <- matrix(rnorm(24), nc = 6)
df <- data.frame(mat)
names(df) <- c("Col1", "Col1", "Col1", "Col2", "Col3", "Col3")

df
rmDuplCols(df)
rmDuplCols(df, keep_first = FALSE)


}
\seealso{
\code{\link{duplicated}}
}
\author{
Florian Detsch
}
