% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrLavaan.R
\name{rrLavaan}
\alias{rrLavaan}
\title{Rubin's Rules for Latent Variable Models}
\usage{
rrLavaan(model, imp.data, type, standardize = FALSE)
}
\arguments{
\item{model}{A user-specified lavaan model}

\item{imp.data}{A multiply-imputed data set of class mids}

\item{type}{Type of lavaan function to call? Either sem, cfa, growth, or lavaan.}

\item{standardize}{Logical. If TRUE, include standardize output. Default is FALSE.}
}
\description{
This function applies Rubin's rules to models fit via lavaan
}
