% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwdOS.R
\name{setwdOS}
\alias{setwdOS}
\title{Set working directory dependent on current OS}
\usage{
setwdOS(lin = "/media/permanent/", win = "D:/", ext = NULL)
}
\arguments{
\item{lin}{Character. Absolute file path to the Linux partition.}

\item{win}{Character. Absolute file path to the Windows partition.}

\item{ext}{Character. Optional file path extension that will be added to
either \code{lin} or \code{win} after OS determination.}
}
\description{
Similar to \code{\link{setwd}}, this function sets the working directory to a
user-defined path. Rather than supplying a single \code{dir} argument, 
however, both an OS-sensitive path to the desired hard disk partition and, 
optionally, an extension of this file path are required.
}
\examples{
\dontrun{
# desired partition
setwdOS()

# including file path extension
setwdOS(ext = "kilimanjaro/nubiscope")
}

}
\seealso{
\code{\link{setwd}}, \code{\link{switch}}
}
\author{
Florian Detsch
}
