% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extend.R
\name{st_extend}
\alias{st_extend}
\title{Extend Method for \strong{sf} Objects}
\usage{
st_extend(x, ...)
}
\arguments{
\item{x}{A \strong{sf} object.}

\item{...}{Arguments passed to \code{\link[sf]{st_buffer}}}
}
\value{
See \code{\link[sf]{st_bbox}}.
}
\description{
Extend simple feature bounding analogous to \strong{raster} 
\code{\link[raster]{extend}} method.
}
\examples{
\dontrun{
data(KiLi)

strs = stars::st_as_stars(KiLi)

## extend by .05 deg
xtd = st_extend(
  strs
  , dist = 0.05
)

## view on map
mapview::viewRGB(as(strs, "Raster"), r = 1, b = 3, quantiles = c(0, 1)) + xtd
}

}
