% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{substrC}
\alias{substrC}
\title{Substrings of a Character Vector (C++ Style)}
\usage{
substrC(x, pos, len)
}
\arguments{
\item{x}{A \code{character} vector.}

\item{pos}{The start point of the substring as \code{integer}. Position 
indications start from \code{1}, which is the default in R.}

\item{len}{The length of the substring as \code{integer}.}
}
\value{
A \code{character} vector of the same length as 'x'.
}
\description{
Extract substrings from a \code{character} vector in C++.
}
\examples{
substrC("Hello, world!", pos = 1, len = 5)

}
\seealso{
\url{http://www.cplusplus.com/reference/string/string/substr/},
\code{\link{substr}}.
}
