% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimImages.R
\name{trimImages}
\alias{trimImages}
\title{Remove whitespace from images}
\usage{
trimImages(path = ".", pttrn = ".png$")
}
\arguments{
\item{path}{Character. Absolute file path to image files.}

\item{pttrn}{Character. A regular expression as taken by
\code{\link{list.files}}; defaults to \code{".png$"}.}
}
\description{
This is a wrapper function around \code{convert -trim} to automatically
remove any whitespace from locally saved images. Note that ImageMagick is
required.
}
\examples{
## trim images in current working directory
trimImages()
}
\author{
Florian Detsch
}
\seealso{
\code{\link{list.files}}, \code{\link{system}}
}

