% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromGit.R
\name{loadFromGit}
\alias{loadFromGit}
\title{Install and load a package from GitHub}
\usage{
loadFromGit(repo = "environmentalinformatics-marburg/Rsenal", ...)
}
\arguments{
\item{repo}{Character. Repository address, see \code{\link{install_github}}. 
Defaults to "environmentalinformatics-marburg/Rsenal".}

\item{...}{Further arguments passed on to \code{\link{install_github}}.}
}
\description{
This function comprises multiple steps required to install and load a 
package directly from GitHub.
}
\examples{
\dontrun{
## install 'Rsenal' package directly from GitHub
loadFromGit()
}

}
\seealso{
\code{\link{install_github}}
}
\author{
Florian Detsch
}
