% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromGit.R
\name{loadFromGit}
\alias{loadFromGit}
\title{Install and Load a Package from GitHub}
\usage{
loadFromGit(repo = "fdetsch/Orcs", ...)
}
\arguments{
\item{repo}{Repository address as \code{character}, defaults to \code{"fdetsch/Orcs"}.}

\item{...}{Additional arguments passed to \code{\link[remotes:install_github]{remotes::install_github()}}.}
}
\description{
This function comprises multiple steps required to install and load a package
directly from GitHub.
}
\examples{
\dontrun{
## install 'Orcs' from GitHub
loadFromGit("fdetsch/Orcs")
}

}
\author{
Florian Detsch
}
