% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPkgs.R
\name{loadPkgs}
\alias{loadPkgs}
\title{Load multiple packages}
\usage{
loadPkgs(pkgs, ...)
}
\arguments{
\item{pkgs}{Packages to load as \code{character}.}

\item{...}{Additional arguments passed to \code{\link{library}}, except for 
'character.only' which is set to \code{TRUE}.}
}
\description{
Load and attach multiple packages at once.
}
\note{
Package startup messages are automatically disabled.
}
\examples{
loadPkgs(c("raster", "rgdal"))

}
\seealso{
\code{\link{library}}.
}
\author{
Florian Detsch
}
