% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orcs-package.R
\docType{package}
\name{Orcs-package}
\alias{Orcs-package}
\alias{orcspackage}
\title{Omnidirectional R Code Snippets.}
\description{
Omnidirectional R Code Snippets
}
\details{
The package provides a variety of functions which I regularly use during my 
everyday work.
}
\author{
Florian Detsch \cr
\cr
\emph{Maintainer:} Florian Detsch \email{florian.detsch@staff.uni-marburg.de}
}
\keyword{package}
