% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OrcsCppFun}
\alias{OrcsCppFun}
\alias{nrowC}
\alias{ncolC}
\alias{dimC}
\title{Dimensions of a \code{data.frame}}
\usage{
nrowC(x)

ncolC(x)

dimC(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
\code{dimC} returns an 'integer' vector of length 2 (number of rows 
and columns); \code{nrowC} (or \code{ncolC}) returns the number of rows 
(or columns) as a single 'integer'.
}
\description{
Similar to base-R \code{\link{nrow}}, \code{\link{ncol}} and 
\code{\link{dim}}, this set of functions let's you retrieve the number of 
rows and columns of a \code{data.frame}.
}
\section{Functions}{
\itemize{
\item \code{nrowC}: 

\item \code{ncolC}: 

\item \code{dimC}: 
}}

\examples{
dat <- data.frame(a = 1:4, b = 2:5, c = 3:6)

nrowC(dat)



}
\seealso{
\code{\link{nrow}}, \code{\link{ncol}}, \code{\link{dim}}.
}
\author{
Florian Detsch
}
