% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pureBasename.R
\name{pureBasename}
\alias{pureBasename}
\title{Return filename without extension}
\usage{
pureBasename(path, slash = FALSE)
}
\arguments{
\item{path}{Filename; see \code{\link{basename}}.}

\item{slash}{If \code{TRUE}, a leading slash ("/") is added to the returned 
filename.}
}
\description{
As opposed to \code{\link{basename}}, this function returns the pure basename 
of a file path, i.e. without extension.
}
\examples{
ch_file <- "path/to/file.with.extension"
pureBasename(ch_file)
pureBasename(ch_file, slash = TRUE)

}
\seealso{
\code{\link{basename}}
}
\author{
Florian Detsch
}
