% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwdOS.R
\name{setwdOS}
\alias{setwdOS}
\title{Set working directory dependent on current OS}
\usage{
setwdOS(path_lin = "/media/permanent/", path_win = "C:/Permanent/",
  path_ext = NULL)
}
\arguments{
\item{path_lin}{Character. Absolute file path to the Linux partition.}

\item{path_win}{Character. Absolute file path to the Windows partition.}

\item{path_ext}{Character. Optional file path extension that will be added to
either \code{path_lin} or \code{path_win} after OS determination.}
}
\description{
Similar to \code{\link{setwd}}, this function sets the working directory to a
user-defined path. Rather than supplying a single \code{dir} argument, 
however, both an OS-sensitive path to the desired hard disk partition and, 
optionally, an extension of this file path are required.
}
\examples{
# desired partition
setwdOS()

# including file path extension
setwdOS(path_ext = "kilimanjaro/nubiscope")

}
\author{
Florian Detsch
}
\seealso{
\code{\link{setwd}}, \code{\link{switch}}
}

